Read in rag with ids not roi in local segmentation
In local_segmentation, fragments in a specific roi are merged based on a rag in specific roi. However, fragment nodes that are at the border of the roi and where the center is outside of the roi are ignored. This commit adresses this problem by reading in the rag based on the fragment ids instead of the ROI.